click and dblclick coord events, and en passant spello
